requirements.txt
# MCP server dependencies
# Install with: pip install -r requirements.txt

fastmcp>=2.0.0
mcp>=1.0.0
uvicorn>=0.20.0
starlette>=0.27.0
sse-starlette>=1.6.0
python-multipart>=0.0.6

# DB
libsql-client>=0.3.0
python-dotenv>=1.0.0

# Data
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pdfplumber>=0.10.0
numpy>=1.24.0
pandas>=2.0.0

# Utils
httpx>=0.24.0
python-dateutil>=2.8.0
